Move some enums out of AbstractPipelineNode into Types.hpp
